Move kprint! macro into its own crate
This allows underlying crates like memory allocators to use the
`kprint!` macro for debug printing to the serial port.
