+++ altos-rust/altos-core/Cargo.toml
@@ -17,6 +17,7 @@ test = []
 [dependencies]
 bump_allocator = { path = "libs/heap/bump_allocator", optional = true }
 volatile = { path = "../libs/volatile" }
+altos-macros = { path = "libs/altos-macros" }
 
 [target.thumbv6m-none-eabi.dependencies]
 cm0_atomic = { path = "libs/cm0_atomic" }
