+++ altos-rust/altos-core/libs/altos-macros/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "altos-macros"
+version = "0.1.0"
+authors = ["Daniel Seitz <dnseitz@gmail.com>"]
+
+[dependencies]
